Commit Inicial

+++ src/main/java/URI_1234/Main.java
@@ -7,7 +7,7 @@ public class Main {
 
     //URI1234
 
-    //Funcionou, mas nao eficientemente
+    //Um pouco ineficiente
 
     public static void main(String[] args) throws IOException {
         Scanner sc = new Scanner(System.in);
@@ -23,20 +23,20 @@ public class Main {
                     output[i] = input.charAt(i);
                 } else if (last == '\0' && input.charAt(i) != ' ') {
                     output[i] = input.toUpperCase().charAt(i);
-                    last = (int) output[i];
+                    last = output[i];
                 } else if (last <= 90 && last >= 65 && input.charAt(i) <= 90 && input.charAt(i) >= 65) {
                     char c = input.charAt(i);
                     int buffer = c + 32;
                     output[i] = (char) buffer;
-                    last = (int) output[i];
+                    last = output[i];
                 } else if (last <= 122 && last >= 97 && input.charAt(i) <= 122 && input.charAt(i) >= 97) {
                     char c = input.charAt(i);
                     int buffer = c - 32;
                     output[i] = (char) buffer;
-                    last = (int) output[i];
+                    last = output[i];
                 } else {
                     output[i] = input.charAt(i);
-                    last = (int) output[i];
+                    last = output[i];
                 }
             }
             System.out.println(new String(output));
